Make release-tfjs update subpackages' link dependencies
