+++ scripts/release-tfjs.ts
@@ -30,6 +30,7 @@ import * as fs from 'fs';
 import * as shell from 'shelljs';
 import {TMP_DIR, $, question, makeReleaseDir, createPR, TFJS_RELEASE_UNIT, updateTFJSDependencyVersions, ALPHA_RELEASE_UNIT, getMinorUpdateVersion, getPatchUpdateVersion, E2E_PHASE, getReleaseBlockers, getNightlyVersion} from './release-util';
 import * as path from 'path';
+import { findDeps } from './graph_utils';
 
 const parser = new argparse.ArgumentParser({
   description: 'Create a release PR for the tfjs monorepo.',
@@ -203,6 +204,7 @@ async function main() {
   // Update versions in package.json files.
   const phases =
       [...TFJS_RELEASE_UNIT.phases, ...ALPHA_RELEASE_UNIT.phases, E2E_PHASE];
+  const errors: Error[] = [];
   for (const phase of phases) {
     for (const packageName of phase.packages) {
       shell.cd(packageName);
@@ -230,27 +232,39 @@ async function main() {
         console.log(chalk.magenta.bold(
             `~~~ Update dependency versions for ${packageJsonPath} ~~~`));
 
-        // Subpackages may depend on the package itself, so add the package to
-        // the list of deps.
-        // TODO(mattSoulanille): Do we need to only update the deps of a phase?
-        // Partial answer: It depends. If the package needs to be built for
-        // release and needs to download from npm in order to build, then if we
-        // update dependencies not listed in the phase's deps, they will not
-        // have been published to npm yet, and the build will fail. On the other
-        // hand, for demos that we don't need to build for publishing to npm,
-        // it's fine to update everything, since they will be run after all
-        // packages are published. Additionally, Bazel packages do not have this
-        // issue since they do not fetch TFJS dependnecies from npm when
-        // building. In general, we should try to move towards updating all
-        // dependencies, or throw an error when it's not possible.
-        const depsWithPackage = [...(phase.deps ?? [])];
+        // // Subpackages may depend on the package itself, so add the package to
+        // // the list of deps.
+        // const depsWithPackage = [...(phase.deps ?? [])];
+        // if (versions.has(packageName)) {
+        //   // Some packages, like e2e, are never published to npm.
+        //   depsWithPackage.push(packageName);
+        // }
+
+        // Only update versions that are a (possibly transitive) dependency of
+        // the package and are listed in the phase deps (we throw an error
+        // if we find a dependency that doesn't satisfy these conditions).
+        const transitiveDeps = [...findDeps([packageName])]
+                                 .filter(dep => phase.deps.includes(dep));
+
+        // Also add the package itself so subpackages can use it.
+        // Some packages, like e2e, are never published to npm, so check first.
         if (versions.has(packageName)) {
-          // Some packages, like e2e, are never published to npm.
-          depsWithPackage.push(packageName);
+          transitiveDeps.push(packageName);
         }
-        const updated = updateTFJSDependencyVersions(pkg, versions,
-                                                     depsWithPackage);
-        fs.writeFileSync(packageJsonPath, updated);
+
+        const packageDependencyVersions = new Map(transitiveDeps.map(dep =>
+          [dep, versions.get(dep)!]));
+
+        try {
+          const updated = updateTFJSDependencyVersions(pkg,
+            packageDependencyVersions);
+
+          fs.writeFileSync(packageJsonPath, updated);
+        } catch (e) {
+          e.message = `For ${packageJsonPath}, ${packageName} ${e.message}`;
+          console.error(e.stack);
+          errors.push(e);
+        }
       }
 
       shell.cd('..');
@@ -261,6 +275,10 @@ async function main() {
       }
     }
   }
+  if (errors.length > 0) {
+    throw new Error('Some package version updates had errors' + errors);
+  }
+
 
   // Use dev prefix to avoid branch being locked.
   const devBranchName = `dev_${releaseBranch}`;
